requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
pandas == 2.3.3
sqlalchemy == 2.0.44
azure-identity==1.17.1
pyodbc==5.1.0
azure-keyvault-secrets==4.10.0
requests==2.32.3
azure-communication-email